Fit Graphical Lasso with RobSel

Fit Graphical Lasso with estimate regularization parameter from Robust Selection

Usage

robsel.glasso(x, alpha = 0.9, B = 200, ...)

Arguments

x

A n-by-p data matrix

alpha

Prespecified confidence level. Default 0.9

B

Number of bootstrap sample. Default 200

...

Optional arguments passed on to glasso.

Value

A list with components:

alpha

A list of prespecified confidence level

lambda

A list of estimate regularization parameter for Graphical Lasso

Omega

A list of estimated inverse covariance matrix

Sigma

A list of estimated covariance matrix

Note

Each item in each compenent corresponds to a prespecified level alpha.

References

P Cisneros-Velarde, A Petersen and S-Y Oh (2020). Distributionally Robust Formulation and Model Selection for the Graphical Lasso. Proceedings of the Twenty Third International Conference on Artificial Intelligence and Statistics.

Friedman, Jerome, Trevor Hastie, and Robert Tibshirani. 'Sparse inverse covariance estimation with the graphical lasso.' Biostatistics 9.3 (2008): 432-441.

Meinshausen, Nicolai and Buhlmann, Peter. 2006. 'High-Dimensional Graphs and Variable Selection with the Lasso.' The Annals of Statistics. JSTOR: 1436-1462.

Witten, Daniela M, Friedman, Jerome H, and Simon, Noah. 2011. 'New Insights and Faster computations for the Graphical Lasso.' Journal of Computation and Graphical Statistics. Taylor and Francis: 892-900.

See also

Examples

set.seed(17)
library(robsel)
x <-matrix(rnorm(50*20),ncol=20)

#Use Graphical Lasso with estimate regularization parameter lambda from RobSel
fit <- robsel.glasso(x = x, alpha = 0.9, B = 200)